Get the Windows build to zero warnings and gate on it
The first green Windows build carried 142 warnings. 134 of them were inside
sqlite3.c: 122 C4232 from SQLite building its table of Win32 API function
pointers, plus a handful of C4127/C4456/C4701. That is vendored third-party code
we do not maintain, and it was drowning out anything of ours, so it is now built
with warnings off -- the same treatment it already gets from gcc via -w.

The remaining 8 were ours: four places in p_hud.c assigning strlen's size_t to
an int without a cast. Fixed rather than suppressed.

The Windows job now fails on warnings the way the Linux job already did, so this
cannot quietly drift back.

+++ p_hud.c
@@ -547,7 +547,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     "xv %i yv %i string2 \"%s\" ",
                     x + 32 - 136 + 80, y, player_rune);
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy(string + stringlength, string2);
@@ -645,7 +645,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     "xv %i yv %i string2 \"R:%s\" ",
                     x + 32 + 80, y + 16, player_rune);
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy(string + stringlength, string2);
@@ -703,7 +703,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     "xv %i yv %i string2 \"%s\" ",
                     x + 32 + 56 + 80, y, player_rune);
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy(string + stringlength, string2);
@@ -805,7 +805,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     "xv %i yv %i string2 \"R:%s\" ",
                     x + 32 + 80, y + 16, player_rune);
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy(string + stringlength, string2);
